+++ addons/l10n_si_guest_feedback/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import feedback_survey
+from . import feedback_response
+from . import feedback_alert
+from . import hotel_folio